Detect conflicting changes on integer domains to stop propagation early
Note that some propagators (like the disjunctive propagator) depend on
the fact that conflicts are detected on integer domains. Because of the
usage of lazy literals, just checking the value of the literals is not
enough. Literals that have just been created might not have their
correct value yet, since their defining clauses still have to be
communicated to the oracle.