Change `SolvingContext::set_bool` to not trail literals early
Changes to the trail now all stem from the `Engine::notify_assignment`
method, called by the SAT oracle. The only exception is the
`Engine::propagate` method, which will call this method early in debug
mode to make literal assignments available when checking that all
literals in the reason are known `true`.